refactor: move render box helpers
