+++ fission/src/util/Utility.ts
@@ -1,7 +1,5 @@
 import type Jolt from "@azaleacolburn/jolt-physics"
-import * as THREE from "three"
 import Pako from "pako"
-import World from "@/systems/World"
 
 export function ternaryOnce<A, B>(obj: A | undefined, ifTrue: (x: A) => B, ifFalse: () => B): B {
     return obj ? ifTrue(obj) : ifFalse()
@@ -72,55 +70,6 @@ export function tryParse(data: string): any {
         console.error("Could not parse JSON", error)
         return null
     }
-}
-
-export function renderAxisAlignedBox(
-    box: Jolt.AABox,
-    color: THREE.ColorRepresentation = 0x00aaff,
-    opacity: number = 0.3
-): THREE.Mesh {
-    const size = box.GetSize()
-    const geo = new THREE.BoxGeometry(size.GetX(), size.GetY(), size.GetZ())
-    const material = new THREE.MeshPhongMaterial({
-        color,
-        opacity,
-        transparent: true,
-        depthWrite: false,
-        side: THREE.DoubleSide,
-    })
-    const mesh = new THREE.Mesh(geo, material)
-
-    const t = box.GetCenter()
-    mesh.position.set(t.GetX(), t.GetY(), t.GetZ())
-
-    World.sceneRenderer.addObject(mesh)
-    return mesh
-}
-
-export function renderOrientedBox(
-    obb: Jolt.OrientedBox,
-    color: THREE.ColorRepresentation = 0x00aaff,
-    opacity: number = 0.3
-): THREE.Mesh {
-    const halfExtents = obb.mHalfExtents
-    const geo = new THREE.BoxGeometry(halfExtents.GetX() * 2, halfExtents.GetY() * 2, halfExtents.GetZ() * 2)
-    const material = new THREE.MeshPhongMaterial({
-        color,
-        opacity,
-        transparent: true,
-        depthWrite: false,
-        side: THREE.DoubleSide,
-    })
-    const mesh = new THREE.Mesh(geo, material)
-
-    const orientation = obb.mOrientation
-    const t = orientation.GetTranslation()
-    const q = orientation.GetQuaternion()
-    mesh.position.set(t.GetX(), t.GetY(), t.GetZ())
-    mesh.quaternion.set(q.GetX(), q.GetY(), q.GetZ(), q.GetW())
-
-    World.sceneRenderer.addObject(mesh)
-    return mesh
 }
 
 export function downloadBlob(filename: string, data: BlobPart): void {
